flask-backend/pyproject.toml
[project]
name = "flask-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Bahaa Shklawon",email = "bahaashklawon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "apiflask (>=2.4.0,<3.0.0)",
    "rembg (>=2.0.67,<3.0.0)",
    "onnxruntime (>=1.22.1,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "dev-kit @ git+https://github.com/bhsh2002/dev-kit.git@v0.1.1"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
